+++ backend/src/production/production.module.ts
@@ -4,6 +4,7 @@ import { ProductionController } from "./controllers/production.controller";
 import { DbModule } from "../database/db.module";
 import { ProductionBlogController } from "./controllers/production-blog.controller";
 import { ProductionTagController } from "./controllers/production-tag.controller";
+import { UtilModule } from "../util/util.module";
 
 @Module({
   providers: [ProductionService],
@@ -12,6 +13,6 @@ import { ProductionTagController } from "./controllers/production-tag.controller
     ProductionBlogController,
     ProductionTagController,
   ],
-  imports: [DbModule],
+  imports: [DbModule, UtilModule],
 })
 export class ProductionModule {}
