+++ backend/src/util/util.module.ts
@@ -1,9 +1,10 @@
 import { Module } from "@nestjs/common";
 import { ScraperService } from "./scraper/scraper.service";
 import { AppLogger } from "./logger/logger.service";
+import { LanguageService } from "./language/language.service";
 
 @Module({
-  providers: [ScraperService, AppLogger],
-  exports: [AppLogger],
+  providers: [ScraperService, AppLogger, LanguageService],
+  exports: [AppLogger, LanguageService],
 })
 export class UtilModule {}
